requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.45
alembic==1.13.2
psycopg2-binary==2.9.9
pandas==2.3.3
passlib[argon2]==1.7.4
argon2-cffi==25.1.0
pyotp==2.9.0
PyJWT==2.10.1
Flask==3.1.2
pytest==9.0.2
tqdm==4.67.1
typing_extensions==4.15.0
pydantic==2.9.2
